Store interfaces that are shared between app and worker elsewhere
