+++ src/common/events.ts
@@ -1,0 +1,4 @@
+export interface SaveFileEventPayload {
+  name: string;
+  file: File;
+}
