+++ src/worker/main.ts
@@ -1,4 +1,5 @@
 import { readExcel } from "./read-wordlist";
+import { SaveFileEventPayload } from "@common/events";
 import Storage from "./storage";
 
 const storage = new Storage();
@@ -6,11 +7,6 @@ const storage = new Storage();
 // Print the function's name (even though that's self-evident) just to get the
 // linter and Rollup to acknowledge that readExcelSync() is used.
 self.postMessage(`hello, from worker! I can use ${readExcel.name}`);
-
-export interface SaveFileEventPayload {
-  name: string;
-  file: File;
-}
 
 const handleSaveFileEvent = async (event: MessageEvent) => {
   const payload = event.data as SaveFileEventPayload;
